+++ apps/web/client/src/app/projects/import/local/_components/finalizing-project.tsx
@@ -5,7 +5,7 @@ import { CardDescription, CardTitle } from '@onlook/ui/card';
 import { ProgressWithInterval } from '@onlook/ui/progress-with-interval';
 import { motion } from 'motion/react';
 import { StepContent, StepFooter, StepHeader } from '../../steps';
-import { useProjectCreation } from '../_context/context';
+import { useProjectCreation } from '../_context';
 
 export const FinalizingProject = () => {
     const { isFinalizing, error, retry, cancel } = useProjectCreation();
